Fix actuator vs switch obstruction
BUTTON_LEVER_SPST_ACTUATOR_CLEARANCE is no longer
needed now that the piece can be printed on its
side.

Previous value was too deep, causing piece to
hit the SPST's base before travel could be
completed.

Leaving it in as 0 instead of removing to
acknowledge that it's known to be unnecessary.

I also played with a smaller cantilever, which
seemed too thick to me. Anything smaller turned
into just one line in the slicer, which was too
brittle.